<feat>커뮤니티 api 연동

+++ FrontEnd/DangDang/app/src/main/java/com/dangdang/common/utils/Utils.kt
@@ -133,8 +133,8 @@ fun addComma(number: Int): String {
     return formattedNumber
 }
 
-fun getMeterToKm(distance: Double): Double{
-    return distance / 1000.0
+fun getMeterToKm(distance: Float): Float{
+    return distance / 1000f
 }
 
 fun navigateBottomTab(
