+++ FrontEnd/DangDang/app/src/main/java/com/dangdang/data/model/chat/AnalysisFoodModel.kt
@@ -4,7 +4,7 @@ import androidx.annotation.Keep
 
 @Keep
 data class AnalysisFoodModel (
-    val predictedGlucoseRise: Double, //예상 혈당 상승량
-    val beginGlucose: Double, //식전 혈당
+    val predictedGlucoseRise: Float, //예상 혈당 상승량
+    val beginGlucose: Float, //식전 혈당
     val foodInfo: FoodInfoModel,
 )
