+++ FrontEnd/DangDang/app/src/main/java/com/dangdang/data/model/walk/WalkMissionEndResponse.kt
@@ -5,8 +5,8 @@ import androidx.annotation.Keep
 @Keep
 data class WalkMissionEndResponse(
     val status: String,
-    val actualDistance: Double,
+    val actualDistance: Float,
     val durationMinutes: Int,
-    val burnedKcal: Double,
+    val burnedKcal: Float,
     val postGlucosePrompted: Boolean
 )
